fix: returning from fn in for loop
These was a bug whereby you could not return within a for loop in a fn:

  fn foo():
      for i in range(5):
          return 5
  a = foo()

This would error because we can't assign void to 'a'... But obviously
that's not what's we're trying to do, it's a bug.

Fix is simple, we were just failing to pass on the eval result from for
loops.

We also test yield stmts while here.