ci: try adding goreleaser & binary releases
I'd like to set up some automatd PR checks (and separately, I've been
wanting to make Linux/Windows installation easier), and to this end,
this commit attempts to add goreleaser for binary releases.

We use goreleaser-cross since Rad has a CGO dependency with tree sitter,
and so we need to use docker to compile for the different platforms.

There will no doubt be some iterations along this path.