feat: allow int short count arg input style
If the user defines an args block like this:

  args:
    num n int

someone invoking this script can now cluster several 'n's together.

  myscript -nn  -> 'num' is 2
  myscript -nnn -> 'num' is 3
  myscript -n   -> 'num' is 1

That last one is potentialy controversial as it makes it easier to make
mistakes. It might be surprising. Perhaps we make counting opt in via an
additional keyword or syntax in the args block.

This commit upgrades us to ra 0.4.0 which includes this functionality
for int64 args (which Rad uses).