ci: implement 5 more PR benchmarks
These target some of the important parts of Rad. If we add many more,
we'll likely wanna change the PR comment to be a summary of all these
benchmarks, rather than printing them each out in their own row.

We also teach our benchmark CI to include benchmarks added in a PR.