feat: validate script immediately after parsing
We currently validate that the script being run is valid a bit late;
after we already try and extract metadata about the script e.g. args
block, etc.

This can lead to unpredictable errors when running invalid scripts.

With this commit, we now validate the input script immediately after
parsing using RadChecker, before attempting to extract args. This:

- Prevents panic by catching invalid syntax early
- Reports clean error messages (respects NO_COLOR env var only)

On that latter point -- we don't try to parse user flags like
'--color=never' if an invalid script is being run. This is for
simplicity, but also with the reasoning that the script *could* be
intended to override the meaning of '--color' anyway, so it potentially
seems arbitrary to interpret --color as the global flag when the script
is invalid, and suddenly not, when the script is fixed.