feat: allow scientific notation for int defaults
Works for both int defaults in args blocks, and int defaults in function
definitions.

Previously, the grammar only allowed scientific notation for floats.
This means you couldn't specify e.g. `num int = 1e2` despite that really
being totally valid.

This was a deliberate choice initially, as the grammar can't easily tell
if a given scientific notation number is actually a valid int or float.
For example, 1.1234e10 is a valid int, 1.1234e2 is not.

In this commit, we update our grammar dependency which separates
scientific notation from ints and floats entirely. Scientific notation
for both ints and float is valid, and gets validated in Rad. This
validation is done in our rad checker logic, which Rad uses on startup,
and is also available via the embedded 'check' command.