fix: correctly allow unknown flags to be absorbed in var args
This was already the intention but it was broken for cases where the
first arg to a var arg was an unknown flag (as opposed to a position
arg).

See test in this commit which tests this scenario.

The fix is entirely in Ra, so this commit is just an upgrade to the
corresponding fix version.