docs: add more commands.md thinking
I am still thinking about the best way to implement commands. For this
commit, I experimented with a 'Rad Co-Owner' output-style for claude
code and discussed some of my ideas with it. We went back and forth a
bit and I think we're coming to something potentially workable for a v1
of the commands feature.

Claude summarized our discussion in existing 'thinking docs' style.