feat: add global config loading
We're going to add some functionality to Rad which controls its behavior
*globally* i.e. in *all* Rad scripts you run on your machine. This
functionality will be opt-in, and we want the ability to configure it.

The best way to accomplish this is via a global config file.

We already have ~/.rad i.e. a RadHome concept.
Let's add a 'config.toml' file in there to control global config.

We pick TOML because it's simple and human-friendly. We can still
programmatically change it (a la uv), if we want to.

It's also getting more and more popular, the fact that the following
examples use it, give me confidence that we can use it in Rad as well
without regretting it: uv, Rust (Cargo), Hugo

This commit just adds the logic and general framework for the global
config, later commits add actual config we have planned.

When Rad starts up, we load our global config. If there are any issues,
we warn to stderr, and we fall back to default config.

We also slightly refactor the Rad exit logic while here, in anticipation
of the next commit, which will implement invocation logging (leveraging
this global config commit).

We use the de-facto standard for TOML reading: BurntSushi/toml