feat: update syntax dump tests for new shell & error handling syntax
We're redoing Rad's shell and error handling syntax. Right now, we have
a 'prefix catch' syntax like this, for non-shell cases:

 a = catch parse_int("two")
 if type_of(a) == "error":
   // handle error ...

This isn't great. On top of that, error handling for shell commands
looks different (and unique):

  $`curl {url}`
  recover:
    // handle error & recover

or you can use the 'unsafe' keyword, or 'fail', or '$!', etc.

It all adds to Rad's learning curve. We want syntax that unifies error
handling for shell and non-shell cases alike, and make it quicker to
handle them in general. We move to a 'catch suffix' approach instead:

  a = parse_int("two") catch:
    // handle error case...

  $`curl {url}` catch:
    // handle error case...

We also add a 'fallback' operator for the common pattern of falling back
to some other value if an operation errors:

  port = parse_int(input_port) ?? 8080

This commit simply updates our tree-sitter-rad dependency and updates
our dump tests with the new grammar, updating the tests as appropriate.