feat: implement new shell syntax + catch suffix syntax
==  Context  ==

We're redoing Rad's shell and error handling syntax. Right now, we have
a 'prefix catch' syntax like this, for non-shell cases:

a = catch parse_int("two")
if type_of(a) == "error":
   // handle error ...

This isn't great. On top of that, error handling for shell commands
looks different (and unique):

  $`curl {url}`
  recover:
    // handle error & recover

or you can use the 'unsafe' keyword, or 'fail', or '$!', etc.

It all adds to Rad's learning curve. We want syntax that unifies error
handling for shell and non-shell cases alike, and make it quicker to
handle them in general. We move to a 'catch suffix' approach instead:

  a = parse_int("two") catch:
    // handle error case...

  $`curl {url}` catch:
    // handle error case...

We also add a 'fallback' operator for the common pattern of falling back
to some other value if an operation errors:

  port = parse_int(input_port) ?? 8080

==  This Commit  ==

This commit implements the new shell syntax & error handling for it. On
top of this, and as part of the new shell syntax, we introduce a new
semantic: named shell output assignment.

Previously to this commit, if you wanted to just get the stderr for a
shell command, you had to first capture (positionally) the code and
stdout:

  code, out, err = $!`echo hi`

You could indicate you're not interested with the first two:

  _, _, err = $!`echo hi`

but this is kinda lame. We can do better. The idea of 'named shell
output assignment' is that, if you assign to vars all matching certain
identifier names, we'll intuit what it is you're after. For our above
example, you can now do this instead:

  stderr = $`echo hi`

Because your var is named exactly 'stderr', Rad will say "okay it's
pretty clear you're after stderr, not code or stdout, so I will capture
just the stderr from this command, and give it to you".

To activate this semantic, all assigned vars must be named one of:

  code
  stdout
  stderr