docs(guide): add type-annotations.md
We had nothing about our Type Annotations system, so let's add.

It's still missing a lot of things. I err on either not mentioning those
things (we can add later), and/or not acknowledging the missing parts
(e.g. Rad doesn't properly validate types in all cases yet). I do this
because the docs will age better this way and be a bit more
maintainable, but also because I will hopefully work on this area more
in the not-so-distant future.