feat: allow non-str types in colorize function
The colorize function currently only allows strings, so you have to
always convert the inputs yourself before invoking. This is a bit
annoying, we should be able to handle e.g. lists of bools.

Let's just accept all types and stringify them. This might not behave
well with odd types such as function references, but eh, that's on the
user. In future we might behave more gracefully there.