fix: allow --version --src and --src-tree on invalid scripts
A previous commit moved forward our script validation -> exit step very
far forward. This meant that, if e.g. --src was invoked on an invalid
script, then we'd exit with the error complaining about e.g. syntax
error.

It's more likely in these cases that the user would like to just see the
src of the script, even if it's invalid. This also allows the nice
workflow of trying to run a script -> getting a syntax error ->
pressing 'up' to get their previous command, and then simply appending
--src to see the full script.

There's no *great* way to implement this without major reworks, so if
the script is invalid, then before we exit, we simply take a quick
detour to check os.Args for the relevant global flags, and execute them
if yes.