fix: un-indent command descriptions
Right now, the indent in command description gets included in the
contents, and thus gets printed with --help.

We can't (or shouldn't) change the grammar to exclude the indent, so
let's change the rts package to unindent for us.

We apply this normalization in a few other places too, though not super
clear it's strictly required.