feat: implement 'filter' rad block field modifier syntax
In a rad block, you can now filter rows by applying a cell-level filter
like so:

  values = [10, 20, 30, 40]
  display:
    fields values
    values:
      filter fn(v) v < 35  // << NEW
      map fn(v) v * 2

This will print out:

  values
  20
  40
  60

With the logic from the original values being as follows:

  10 -> 10 < 35 PASS -> *2 -> 20
  20 -> 20 < 35 PASS -> *2 -> 40
  30 -> 30 < 35 PASS -> *2 -> 60
  40 -> 40 < 35 FAIL -> NOT PRINTED

i.e. filtering occurs *before* mapping.

We may want to modify the syntax slightly in future to allow it to
access to e.g. the cell's index, etc. Though this also applies to map.

In this change, while in this area of the code, we also enforce a new
semantic for field modifiers:

- rad/request blocks: modifiers permanently mutate the fields
- display block: modifiers are temporary for *just* the display block

It remains to be seen if this ends up confusing or not, we may revisit.
The logic is that rad/request runs the actual JSON field extraction algo
and is *already* mutating incoming fields, where the display block has 0
side effects and just presents the table of input values.