feat: add flat_map function
Spawned from discussion in #71, I realized we're missing a flattening
function, which is quite often handy to have.

We call it flat_map() and make it operate on lists or maps.
If used on lists, then providing an actual mapping function isn't
required. Instead, identify will be assumed i.e. you don't need to write

  my_list.flat_map(fn(e) e)

as you do in other languages like Java.

However, for *do* require the function to be defined inputting map
types, because we can't assume identity in a way that makes sense (given
key-value pairs).