feat(pick): implement 'prefer_exact' named arg
Previously, if a filter had an exact match on one entry, and a partial
match on another entry, then we'd still do an interactive select between
the two.

This might be preferred in some cases, especially pick and pick_kv as,
based on my current usage thus far, the input data to these are often
runtime resolved rather than pre-planned.

That's not true for pick_from_resource however. The keys defined in
resources (in my experience thus far) are deliberately chosen, and so
we'd actually prefer to just immediately pick an entry if there's a
single exact match.

Because of these preferred defaults, but also because you might not
*always* want those defaults, we introduce the 'prefer_exact' bool named
arg to control this behavior. For pick and pick_kv, it's default false,
but it's default true for pick_from_resource.

This diff in behavior between the 2 and 1 version of the pick functions
*might* be a little surprising, but I think having those respective
default behaviors out of the box outweights that concern.